scripts: implement mode 9 - ksu_susfs_bind_kstat
this is actually a legacy way of handling susfs hide, since very few users have the
need for this, better have this opt-in for them instead. I'm just getting tired
of "I need kstat mode 1", "kstat on mode 1 fucks my setup". this should solve it.

since the ones needing kstat are what I consider broken, they are the ones to adapt.
ending this bullshit once and for all.

A follow up on:
	https://github.com/bindhosts/bindhosts/commit/b7f05e644c8c3e7226217a52732ee1ce1ce2f14c
	https://github.com/bindhosts/bindhosts/commit/425bf05247ab18e07f8ff3e23f07f59903bacf8f
	https://github.com/bindhosts/bindhosts/commit/ef5060ea70499e48deb1fef578201704e3f7fd4d
	https://github.com/bindhosts/bindhosts/commit/c5900e2bbbf10aea46193ba2d6be9443c30f3bcd
	https://github.com/bindhosts/bindhosts/commit/28ddeade34de7d685ad77603e961aac6827ad5e5
	https://github.com/bindhosts/bindhosts/commit/5fe23991876fa77e3454bf43a6c2732fcea08cfd

Signed-off-by: backslashxx <118538522+backslashxx@users.noreply.github.com>